fix ambiguous casting

+++ Arduino/libraries/Adafruit_ADXL345/Adafruit_ADXL345_U.cpp
@@ -104,7 +104,7 @@ uint8_t Adafruit_ADXL345_Unified::readRegister(uint8_t reg) {
     Wire.beginTransmission((uint8_t)_i2caddr);
     i2cwrite(reg);
     Wire.endTransmission();
-    Wire.requestFrom((uint8_t)_i2caddr, 1);
+    Wire.requestFrom((uint8_t)_i2caddr, (uint8_t)1);
     return (i2cread());
   } else {
     reg |= 0x80; // read byte
@@ -126,7 +126,7 @@ int16_t Adafruit_ADXL345_Unified::read16(uint8_t reg) {
     Wire.beginTransmission((uint8_t)_i2caddr);
     i2cwrite(reg);
     Wire.endTransmission();
-    Wire.requestFrom((uint8_t)_i2caddr, 2);
+    Wire.requestFrom((uint8_t)_i2caddr, (uint8_t)2);
     return (uint16_t)(i2cread() | (i2cread() << 8));  
   } else {
     reg |= 0x80 | 0x40; // read byte | multibyte
